カラータイルの結果検証をCLI経由で実施 + その結果をDBに保存

+++ lib/elixirika/tile_score.ex
@@ -8,7 +8,7 @@ defmodule Elixirika.TileScore do
     field :username,    :string,  default: "",    meta: [null: false]
     field :playlog,     :string,  default: "",    meta: [null: false, size: 15000] # TODO TEXT型にしたい
     field :difficulty,  :integer, default: 0,     meta: [null: false]
-    field :remain_time, :integer, default: 0,     meta: [null: false]
+    field :remain_time, :float,   default: 0.0,   meta: [null: false]
     field :extinct,     :boolean, default: false, meta: [null: false]    
   end
 end
